chore: update @wendellhu/redi to version 0.18.0
