+++ packages/design/src/components/confirm/Confirm.tsx
@@ -16,17 +16,10 @@
 
 import type { ILocale } from '../../locale/interface';
 
-import React, { useContext, useRef } from 'react';
+import React, { useContext } from 'react';
 import { Button } from '../button/Button';
 import { ConfigContext } from '../config-provider/ConfigProvider';
 import { Dialog } from '../dialog/Dialog';
-
-export interface IConfirmChildrenProps {
-    hooks: {
-        beforeConfirm?: () => Record<string, any>;
-        beforeClose?: () => Record<string, any>;
-    };
-}
 
 export interface IConfirmProps {
     children: React.ReactNode;
@@ -55,12 +48,12 @@ export interface IConfirmProps {
     /**
      * Callback when the Confirm is closed.
      */
-    onClose?: (result?: Record<string, any>) => void;
+    onClose?: () => void;
 
     /**
      * Callback when the Confirm is confirmed.
      */
-    onConfirm?: (result?: Record<string, any>) => void;
+    onConfirm?: () => void;
 
     /**
      * The width of the Confirm.
@@ -84,17 +77,7 @@ function Footer(props: { locale: ILocale['design']; cancelText?: string; confirm
 
 export function Confirm(props: IConfirmProps) {
     const { children, visible = false, title, cancelText, confirmText, width, onClose, onConfirm } = props;
-
     const { locale } = useContext(ConfigContext);
-    const hooks = useRef<IConfirmChildrenProps['hooks']>({});
-
-    const childrenWithHooks = React.Children.map(children, (child) => {
-        if (React.isValidElement(child)) {
-            return React.cloneElement(child, { hooks: hooks.current });
-        }
-
-        return child;
-    });
 
     return (
         <Dialog
@@ -106,38 +89,14 @@ export function Confirm(props: IConfirmProps) {
                     locale={locale!}
                     cancelText={cancelText}
                     confirmText={confirmText}
-                    onClose={() => {
-                        const beforeClose = hooks.current.beforeClose;
-                        if (beforeClose) {
-                            const result = beforeClose();
-                            if (result.cancel) {
-                                return;
-                            }
-                            onClose?.(result);
-                            return;
-                        }
-
-                        onClose?.();
-                    }}
-                    onConfirm={() => {
-                        const beforeConfirm = hooks.current.beforeConfirm;
-                        if (beforeConfirm) {
-                            const result = beforeConfirm();
-                            if (result.cancel) {
-                                return;
-                            }
-                            onConfirm?.(result);
-                            return;
-                        }
-
-                        onConfirm?.();
-                    }}
+                    onClose={onClose}
+                    onConfirm={onConfirm}
                 />
             )}
             onClose={onClose}
             width={width}
         >
-            {childrenWithHooks}
+            {children}
         </Dialog>
     );
 }
