+++ packages/ui/package.json
@@ -84,7 +84,7 @@
         "@univerjs/design": "workspace:*",
         "@univerjs/engine-render": "workspace:*",
         "@univerjs/icons": "^0.3.22",
-        "@wendellhu/redi": "0.17.1",
+        "@wendellhu/redi": "0.18.0",
         "localforage": "^1.10.0",
         "rc-notification": "^5.6.2"
     },
