+++ packages/ui/src/views/components/confirm-part/interface.ts
@@ -18,8 +18,21 @@ import type { IConfirmProps } from '@univerjs/design';
 
 import type { ICustomLabelProps } from '../../../components/custom-label/CustomLabel';
 
+export interface IConfirmChildrenProps {
+    hooks: {
+        beforeConfirm?: () => Record<string, any>;
+        beforeClose?: () => Record<string, any>;
+    };
+}
+
+export interface IContextConfirmProps extends IConfirmProps {
+    children: React.ReactElement<IConfirmChildrenProps>;
+    onClose?: (result?: Record<string, any>) => void;
+    onConfirm?: (result?: Record<string, any>) => void;
+}
+
 export type IConfirmPartMethodOptions = {
     id: string;
     children?: ICustomLabelProps | string;
     title?: ICustomLabelProps;
-} & Omit<IConfirmProps, 'children' | 'title'>;
+} & Omit<IContextConfirmProps, 'children' | 'title'>;
